// TAA_HTML/src/types/taa.ts
export type ReviewStatus = 'NOT_TOUCHED' | 'PENDING' | 'REVIEWED';